Move normalise_query_string into wagtailsearch utils and use it in the backends
