+++ wagtail/wagtailsearch/views/queries.py
@@ -6,6 +6,7 @@ from wagtail.wagtailadmin.modal_workflow import render_modal_workflow
 from wagtail.wagtailadmin.forms import SearchForm
 
 from wagtail.wagtailsearch import models
+from wagtail.wagtailsearch.utils import normalise_query_string
 
 
 @permission_required('wagtailadmin.access_admin')
@@ -19,7 +20,7 @@ def chooser(request, get_results=False):
         searchform = SearchForm(request.GET)
         if searchform.is_valid():
             query_string = searchform.cleaned_data['q']
-            queries = queries.filter(query_string__icontains=models.Query.normalise_query_string(query_string))
+            queries = queries.filter(query_string__icontains=normalise_query_string(query_string))
     else:
         searchform = SearchForm()
